third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/style-query-registered-custom-em-change-expected.txt

This is a testharness.js-based test.
[FAIL] Changing the container's font-size to 10px makes 1em (10px) + 10px evaluate to 20px. Target matches.
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
